Refactored S2 cells to use the new Polygon class.
