+++ lib/features/nest.js
@@ -10,6 +10,7 @@ const BASE_URL = location.protocol + '//' + location.host + location.pathname;
 class Nest extends Polygon {
   constructor(options) {
     super(options);
+    this.cssId = this.id.toLowerCase().replace(/\//g, '');
     this.pokemon = options.pokemon;
     this.permalinkName = this.name.replace(/[^\w]/g, '').toLowerCase();
     this.dateAdded = options.dateAdded;
