+++ lib/features/polygon.js
@@ -7,7 +7,6 @@ class Polygon {
   constructor(options) {
     this.options = options;
     this.id = options.id;
-    this.cssId = this.id.toLowerCase().replace(/\//g, '');
     this.name = options.name;
     this.region =
       $.map(options.coords, coord => Utils.coordinateToLatLng(coord));
