+++ lib/map.js
@@ -6,7 +6,7 @@ const SearchBox = require('./search_box');
 const Gym = require('./features/gym');
 const Pokestop = require('./features/pokestop');
 const Portal = require('./features/portal');
-const S2Cell = require('./s2_cell');
+const S2Cell = require('./features/s2_cell');
 
 window.initMap = () => {
   const latLngStr = Utils.urlParameter('ll');
@@ -131,7 +131,10 @@ function initS2Cells(map) {
       cellLevelLimit = level;
     }
 
-    if (Utils.isMobile() && (cellLevelLimit > 0)) cellLevelLimit -= 1;
+    if (Utils.isMobile() && (cellLevelLimit > 0)) {
+      cellLevelLimit -= 1;
+    }
+
     return map.getZoom() >= cellLevelLimit;
   };
 
@@ -367,13 +370,9 @@ function loadAndDrawFeatureDataIncrementally(map) {
             $.each(featureData, (index, feature) => {
               if (feature.shouldShow()) {
                 feature.show(map);
-
-                if (!loader.removed) {
-                  loader.remove();
-                  loader.removed = true;
-                }
               }
             });
+
             return featureData;
           });
         featureDataPromises.push(promise);
@@ -381,11 +380,14 @@ function loadAndDrawFeatureDataIncrementally(map) {
 
       return Promise.all(featureDataPromises);
     })
-    .then(featureDataChunks =>
-      featureDataChunks.reduce((featureData, chunk) => 
-        featureData.concat(chunk), []
+    .then(featureDataChunks => {
+      loader.remove();
+
+      return featureDataChunks.reduce(
+        (featureData, chunk) => featureData.concat(chunk)
+        , []
       )
-    )
+    })
     .then(featureData => {
       let dateOfLastUpdate = new Date(1990, 0, 1);
       $.each(featureData, (index, feature) => {
